+++ src/hiero_sdk_python/__init__.py
@@ -110,6 +110,8 @@ from .contract.ethereum_transaction import EthereumTransaction
 # Schedule
 from .schedule.schedule_create_transaction import ScheduleCreateTransaction
 from .schedule.schedule_id import ScheduleId
+from .schedule.schedule_info import ScheduleInfo
+from .schedule.schedule_info_query import ScheduleInfoQuery
 
 # Nodes
 from .nodes.node_create_transaction import NodeCreateTransaction
@@ -223,6 +225,8 @@ __all__ = [
     # Schedule
     "ScheduleCreateTransaction",
     "ScheduleId",
+    "ScheduleInfoQuery",
+    "ScheduleInfo",
 
     # Nodes
     "NodeCreateTransaction",